projects/unusual-coloring-book/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
google-genai>=1.0.0
pillow>=10.4.0
pydantic>=2.9.0
python-multipart>=0.0.9
aiohttp>=3.10.0
